pyproject.toml
[tool.poetry]
name = "stable-diffusion-discord-bot"
version = "0.1.0"
description = ""
authors = ["hommevert <126019798+hommevert@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"
replicate = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
